Do not call task.execute()
Either call the TaskAction directly or move to integration tests
