+++ src/test/groovy/com/netflix/gradle/plugins/docker/OsPackageDockerPluginTest.groovy
@@ -60,7 +60,7 @@ class OsPackageDockerPluginTest extends ProjectSpec {
             instruction 'WORKDIR /tmp'
         }
 
-        task.execute()
+        task.copy()
 
         then:
         task.archivePath.exists()
