+++ src/test/groovy/com/netflix/gradle/plugins/packaging/SystemPackagingPluginTest.groovy
@@ -2,15 +2,12 @@ package com.netflix.gradle.plugins.packaging
 
 import com.netflix.gradle.plugins.deb.Deb
 import com.netflix.gradle.plugins.rpm.Rpm
-import com.netflix.gradle.plugins.rpm.Scanner
 import nebula.test.ProjectSpec
 import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.testfixtures.ProjectBuilder
 import spock.lang.Issue
 import spock.lang.Unroll
-
-import static org.redline_rpm.header.Header.HeaderTag.DESCRIPTION
 
 class SystemPackagingPluginTest extends ProjectSpec {
     def 'tasks created'() {
@@ -66,57 +63,5 @@ class SystemPackagingPluginTest extends ProjectSpec {
         then: "Task should have output"
         debTask.outputs.hasOutput
 
-    }
-
-    @Issue("https://github.com/nebula-plugins/gradle-ospackage-plugin/issues/104")
-    @Unroll
-    def "Translates extension packageDescription '#description' to header entry for RPM task"() {
-        given:
-        project.apply plugin: 'nebula.ospackage'
-
-        project.ospackage {
-            packageName = 'bleah'
-            packageDescription = description
-            version = '1.0'
-        }
-
-        when:
-        project.tasks.getByName('buildRpm').execute()
-
-        then:
-        def scan = Scanner.scan(project.file('build/distributions/bleah-1.0.noarch.rpm'))
-        Scanner.getHeaderEntryString(scan, DESCRIPTION) == headerEntry
-
-        where:
-        description             | headerEntry
-        'This is a description' | 'This is a description'
-        ''                      | ''
-        null                    | ''
-    }
-
-    @Issue("https://github.com/nebula-plugins/gradle-ospackage-plugin/issues/104")
-    @Unroll
-    def "Translates extension packageDescription '#description' to header entry for Debian task"() {
-        given:
-        project.apply plugin: 'nebula.ospackage'
-
-        project.ospackage {
-            packageName = 'translates-extension-description'
-            packageDescription = description
-        }
-
-        when:
-        Task buildDep = project.tasks.getByName('buildDeb')
-        buildDep.execute()
-
-        then:
-        def scan = new com.netflix.gradle.plugins.deb.Scanner(buildDep.archivePath)
-        scan.getHeaderEntry('Description') == headerEntry
-
-        where:
-        description             | headerEntry
-        'This is a description' | 'translates-extension-description\n This is a description'
-        ''                      | 'translates-extension-description'
-        null                    | 'translates-extension-description'
     }
 }
